pyproject.toml
[project]
name = "media-viewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "jinja2>=3.1.6",
    "pillow>=12.1.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.42.0",
]
